# webconsole-back/swagger/SGDSI_Website4sgCore_API_ADMIN_PROD.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Website4sg Core API ADMIN PROD documentation",
    "version" : "1.3",
    "title" : "SGDSI Website4sg Core API ADMIN PROD"
  },
  "host" : "core.website4sg.saint-gobain.com",
  "basePath" : "/",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/v1/accessrights" : {
      "get" : {
        "tags" : [ "v1accessrights" ],
        "summary" : "Get all Access rights managed by Website4sg",
        "description" : "By default, fetchLimit parameter is set to 25",
        "operationId" : "getAllAccessRightBody",
        "parameters" : [ {
          "name" : "email",
          "in" : "query",
          "description" : "The user email",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Results page you want to retrieve (0..N), default value = O",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page, default value = 20",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AccessRightBody"
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/v1/accessrights/{accessRightId}" : {
      "get" : {
        "tags" : [ "v1accessrights" ],
        "summary" : "Get the AccessRight",
        "description" : "",
        "operationId" : "getAccessRightBody",
        "parameters" : [ {
          "name" : "accessRightId",
          "in" : "path",
          "description" : "AccessRight Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "$ref" : "#/definitions/AccessRightBody"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "v1accessrights" ],
        "summary" : "Create or update an accessRight to a User for a Project with a Role",
        "description" : "",
        "operationId" : "createOrUpdateAccessRight",
        "parameters" : [ {
          "name" : "accessRightId",
          "in" : "path",
          "description" : "AccessRight Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "AccessRight detail",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/AccessRightBody"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully updated",
            "schema" : {
              "$ref" : "#/definitions/AccessRightBody"
            }
          },
          "201" : {
            "description" : "Successfully created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v1accessrights" ],
        "summary" : "Remove an accessRight",
        "description" : "",
        "operationId" : "deleteEnvironmentDetail",
        "parameters" : [ {
          "name" : "accessRightId",
          "in" : "path",
          "description" : "AccessRight Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully deleted"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
    "/v1/hosting/docroots/{docrootCode}/env/{environmentCode}" : {
      "get" : {
        "tags" : [ "v1hostingdocroots" ],
        "summary" : "Get the environment detail, associated to the specified docroot",
        "description" : "",
        "operationId" : "getEnvironmentDetail",
        "parameters" : [ {
          "name" : "docrootCode",
          "in" : "path",
          "description" : "Docroot Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "environmentCode",
          "in" : "path",
          "description" : "Environment Id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "$ref" : "#/definitions/DocrootEnvironmentDetail"
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "v1hostingdocroots" ],
        "summary" : "Create Or Update an environment on the specified docroot",
        "description" : "",
        "operationId" : "createOrUpdateDocrootEnvironment",
        "parameters" : [ {
          "name" : "docrootCode",
          "in" : "path",
          "description" : "Docroot Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "environmentCode",
          "in" : "path",
          "description" : "Environment Id",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "DocrootEnvironement detail",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/DocrootEnvironmentDetail"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully updated",
            "schema" : {
              "$ref" : "#/definitions/DocrootEnvironmentDetail"
            }
          },
          "201" : {
            "description" : "Successfully created"
          },
          "400" : {
            "description" : "The body is empty"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v1hostingdocroots" ],
        "summary" : "Remove the environement from the docroot",
        "description" : "",
        "operationId" : "deleteEnvironmentDetail",
        "parameters" : [ {
          "name" : "docrootCode",
          "in" : "path",
          "description" : "Docroot Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "environmentCode",
          "in" : "path",
          "description" : "Environment Id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully deleted"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
  },
  "definitions" : {
    "AccessRightBody" : {
      "type" : "object",
      "required" : [ "projectCode", "projectType", "roleLabel", "userMail" ],
      "properties" : {
        "accessRightId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "userMail" : {
          "type" : "string"
        },
        "roleLabel" : {
          "type" : "string"
        },
        "projectCode" : {
          "type" : "string"
        },
        "projectType" : {
          "type" : "string",
          "description" : "Project Type",
          "enum" : [ "ddc", "w" ]
        },
        "projectName" : {
          "type" : "string"
        }
      },
      "description" : "AccessRightBody Object"
    },
...
    "DocrootEnvironmentDetail" : {
      "type" : "object",
      "required" : [ "cmsCode", "drupalDocrootCoreCode" ],
      "properties" : {
        "cmsCode" : {
          "type" : "string"
        },
        "cmsVersion" : {
          "type" : "string"
        },
        "drupalDocrootCoreCode" : {
          "type" : "string"
        },
        "drupalDocrootCoreVersion" : {
          "type" : "string"
        },
        "canAutoDeploy" : {
          "type" : "boolean"
        },
        "providerInternalId" : {
          "type" : "string"
        },
        "acquiaEnvironmentId" : {
          "type" : "string"
        }
      },
      "description" : "DocrootEnvironmentDetail Object"
    },
...
  }
}